Added a Windows start up script.
If Python is correctly installed on Windows, 'python setup.py install'
will copy the startup script to the Python script directory, which
should be in the PATH environment variable. This means typing 'photini'
in a command shell will start the metadata editor. (You can also double
click on the batch file, if that's what you prefer.)