Improved thumbnail drag & drop behaviour.
I had set it so only selected images could be dragged, and images were
only selected when the mouse button was released. It's now a bit more
sensible - clicking on a thumbnail selects it (control or shift allow
selecting multiple images) and you can immediately drag the selected
image(s).