Created 'unselected' markers for Bing and OSM maps.
Unlike Google maps these providers don't supply a set of icons, so I had
to create my own plain grey icons of the appropriate size and shape to
use for unselected photos.
There is still one small problem: unselected markers should not be
draggable, but this is not the case on the OSM map.