Complete rework of date/time metadata handling
Instead of storing separate date and time, which can be None to allow
some measure of variable precision, we now use a single datetime object
with a 'precision' value that indicates how many fields are relevant.
This allows users to set just a year for archive photos with uncertain
dates. It also simplifies the technical tab date & time widgets.