Substantial rework of metadata module
There is now one big list of tags, each of which has a mode to indicate
if and when it is read, written or cleared. This avoids other parts
having to infer things from the tag names.

This big change may have broken things that I haven't found yet. Use it
with caution.