Pixelfed uploader can replace alt text on existing
This is a bit unintuitive, as one "upload" is a group of images, but
any image having its descriptive text changed is removed from the
group. In all the uploaders, modifying a previous upload isn't really
the same as doing an upload (even if the api calls are mostly the same)
so probably ought to be a different user operation.