Select images opened on remote command
If you use the "open with Photini" context menu option in an image
viewer application or system file browser then Photini will now select
the image(s) that have been requested. This should make it easier to
find the image you want to edit in a viewer, then switch to Photini to
edit it. This idea came from discussion in enhancement request #268.