Save keyword model on quit
Previously the keyword model was only changed by the tree view dialog.
Now it's updated whenever an image's keywords are shown and is saved
when the tree view is opened, to allow "cancel" to revert to the state
when the dialog was opened.

I didn't fancy trying to explain how it worked previously in the docs.
If you can't explain it clearly then it's probably wrong.