Handle missing namespace info for 'apple-fi'
Image region data from some Apple phones includes face recognition
data with an Xmp prefix of 'apple-fi'. This isn't a top level prefix so
exiv2 doesn't find it when it uses the XMP toolkit to parse a file.

This change registers the unknown namespace (with a dummy URL) so the
image can at least be loaded and saved. This fixes bug #325.